refactor: rename permission context helper
